+++ src/pc/constants.js
@@ -8,8 +8,10 @@ const packageJSON = require('./package.json')
 
 const PLATFORM = os.platform()
 const ZADARK_VERSION = packageJSON.version
+const IS_MAC = PLATFORM === 'darwin'
 
 module.exports = {
   PLATFORM,
-  ZADARK_VERSION
+  ZADARK_VERSION,
+  IS_MAC
 }